multiple changes - rename top level folder - add color and better icons - add extra spacing to the board, and each icon keeps track of its size
