--- build_a_game/part-5-grid-stop/player_character.rb
@@ -1,12 +1,0 @@
-class PlayerCharacter
-  ICON = "@"
-
-  attr_accessor :location_x, :location_y, :icon
-
-  def initialize(location_x:, location_y:)
-    @location_x = location_x.to_i
-    @location_y = location_y.to_i
-
-    @icon = ICON
-  end
-end
