+++ build_a_game/part-5-grid_stop/coin.rb
@@ -1,0 +1,19 @@
+require "./thing.rb"
+
+# 2022
+# 2092 - this one is good
+# 2296 or 2299 or 229D - good, easier to read
+# 25CB - circle
+class Coin < Thing
+  COLOR = "93m"
+  ICON = " " + ESCAPE + COLOR + "\u25CB" + CLEAR_CODE + " "
+
+  attr_accessor :location_x, :location_y, :icon
+
+  def initialize(location_x:, location_y:)
+    @location_x = location_x.to_i
+    @location_y = location_y.to_i
+
+    @icon = ICON
+  end
+end
