+++ build_a_game/part-5-grid_stop/game.rb
