+++ build_a_game/part-5-grid_stop/thing.rb
@@ -1,0 +1,4 @@
+class Thing
+  ESCAPE = "\e["
+  CLEAR_CODE = "\e[0m"
+end
